LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A1_T4-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Cannot use the undeclared label 'LABEL'.
S12.8_A1_T4

PASS Expected parsing failure

TEST COMPLETE
